Module rustbreak::deser

Different serialization and deserialization methods one can use

Structs

Bincode
The struct that allows you to use bincode

Ron
The Struct that allows you to use ron the Rusty Object Notation

Yaml
The struct that allows you to use yaml

Traits

DeSerializer
A trait to bundle serializer and deserializer in a simple struct